# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenDeX Backend API",
    "version": "0.1.0",
    "summary": "Local control API of the OpenDeX backend (Android multi-window mirroring).",
...
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "contact": {
      "name": "OpenDeX on GitHub",
      "url": "https://github.com/Lordof21/OpenDex"
    }
  },
  "paths": {
...
    "/api/v1/pairing/qr": {
      "post": {
        "tags": [
          "pairing"
        ],
        "summary": "Create Pairing Qr",
        "description": "Generates a secure QR payload and spawns an mDNS TLS pairing listener.",
        "operationId": "create_pairing_qr_api_v1_pairing_qr_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QrPayload"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/pairing/manual": {
      "post": {
        "tags": [
          "pairing"
        ],
        "summary": "Pair Manual",
        "description": "mDNS güvenlik ağı — hotspot multicast'i güvenilmez olduğunda elle IP:port.",
        "operationId": "pair_manual_api_v1_pairing_manual_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/pairing/detected-ip": {
      "get": {
        "tags": [
          "pairing"
        ],
        "summary": "Get Detected Ip",
        "description": "Detects the phone's IP directly from an attached device or Windows gateway.",
        "operationId": "get_detected_ip_api_v1_pairing_detected_ip_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          }
        }
      }
    },
    "/api/v1/pairing/pair-code": {
      "post": {
        "tags": [
          "pairing"
        ],
        "summary": "Pair With Code",
        "description": "mDNS bağımsız eşleştirme — QR akışı `_adb-tls-pairing._tcp` yayınını hiç\nalmadığında (güvenlik duvarı/VPN/hotspot multicast engeli) kullanılan yedek.",
        "operationId": "pair_with_code_api_v1_pairing_pair_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualPairCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings Endpoint",
        "description": "Retrieves current global project settings.",
        "operationId": "get_settings_endpoint_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettings"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Settings",
        "description": "Updates settings and applies live quality/audio reconfigurations.",
        "operationId": "put_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
    "/api/v1/layout": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Layout",
        "description": "Retrieves saved desktop icon grid layout for current device.",
        "operationId": "get_layout_api_v1_layout_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppLayoutEntry"
                  },
                  "type": "array",
                  "title": "Response Get Layout Api V1 Layout Get"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Layout",
        "description": "Persists desktop icon positions for current device.",
        "operationId": "put_layout_api_v1_layout_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/AppLayoutEntry"
                },
                "type": "array",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AppLayoutEntry"
                  },
                  "type": "array",
                  "title": "Response Put Layout Api V1 Layout Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/ForeignOrigin"
          },
          "411": {
            "$ref": "#/components/responses/LengthRequired"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppLayoutEntry": {
        "properties": {
          "package": {
            "type": "string",
            "title": "Package"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "package",
          "position"
        ],
        "title": "AppLayoutEntry"
      },
...
      "DeviceInfo": {
        "properties": {
          "serial": {
            "type": "string",
            "title": "Serial"
          },
          "state": {
            "$ref": "#/components/schemas/DeviceState"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "transport": {
            "type": "string",
            "enum": [
              "usb",
              "wireless"
            ],
            "title": "Transport",
            "default": "usb"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": false
          },
          "transport_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transport Id"
          }
        },
        "type": "object",
        "required": [
          "serial",
          "state"
        ],
        "title": "DeviceInfo"
      },
...
      "DeviceState": {
        "type": "string",
        "enum": [
          "device",
          "unauthorized",
          "offline",
          "unknown"
        ],
        "title": "DeviceState"
      },
...
      "ErrorDetail": {
        "title": "ErrorDetail",
        "type": "object",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "array"
              },
              {
                "type": "object"
              }
            ],
            "description": "A human-readable message, a machine code (e.g. `not_supported`), or — for 422 — the list of validation errors."
          }
        },
        "required": [
          "detail"
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualPairCodeRequest": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port"
          },
          "pairing_code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^\\d{6}$",
            "title": "Pairing Code"
          }
        },
        "type": "object",
        "required": [
          "ip",
          "port",
          "pairing_code"
        ],
        "title": "ManualPairCodeRequest"
      },
      "ManualPairRequest": {
        "properties": {
          "ip": {
            "type": "string",
            "title": "Ip"
          },
          "port": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Port"
          }
        },
        "type": "object",
        "required": [
          "ip",
          "port"
        ],
        "title": "ManualPairRequest"
      },
...
      "ProjectSettings": {
        "properties": {
          "dynamic_resolution_enabled": {
            "type": "boolean",
            "title": "Dynamic Resolution Enabled",
            "default": false
          },
          "resolution_mode": {
            "type": "string",
            "enum": [
              "dynamic",
              "dynamic_fit",
              "dynamic_fix",
              "phone_scale",
              "1080p",
              "tablet",
              "2k",
              "2.5k",
              "fixed_1080p",
              "fixed_1200p",
              "fixed_1440p",
              "fixed_1600p"
            ],
            "title": "Resolution Mode",
            "default": "dynamic"
          },
          "video_fit_mode": {
            "type": "string",
            "enum": [
              "contain",
              "fill",
              "cover"
            ],
            "title": "Video Fit Mode",
            "default": "contain"
          },
          "screen_off_while_mirroring": {
            "type": "boolean",
            "title": "Screen Off While Mirroring",
            "default": false
          },
          "enable_audio": {
            "type": "boolean",
            "title": "Enable Audio",
            "default": true
          },
          "max_fps": {
            "type": "integer",
            "maximum": 240.0,
            "minimum": 1.0,
            "title": "Max Fps",
            "default": 60
          },
          "video_bit_rate": {
            "type": "integer",
            "maximum": 500000000.0,
            "minimum": 1.0,
            "title": "Video Bit Rate",
            "default": 8000000
          },
          "max_size": {
            "type": "integer",
            "maximum": 8192.0,
            "minimum": 0.0,
            "title": "Max Size",
            "default": 1280
          },
          "custom_dpi": {
            "type": "integer",
            "maximum": 1200.0,
            "minimum": 0.0,
            "title": "Custom Dpi",
            "default": 0
          },
          "target_dp": {
            "type": "integer",
            "maximum": 4000.0,
            "minimum": 0.0,
            "title": "Target Dp",
            "default": 0
          },
          "ambient_backdrop": {
            "type": "boolean",
            "title": "Ambient Backdrop",
            "default": true
          },
          "enable_hybrid_dom": {
            "type": "boolean",
            "title": "Enable Hybrid Dom",
            "default": false
          },
          "audio_codec": {
            "type": "string",
            "enum": [
              "raw",
              "opus"
            ],
            "title": "Audio Codec",
            "default": "raw"
          },
          "custom_encoder_limit": {
            "type": "integer",
            "maximum": 64.0,
            "minimum": 0.0,
            "title": "Custom Encoder Limit",
            "default": 0
          },
          "header_hover_mode": {
            "type": "boolean",
            "title": "Header Hover Mode",
            "default": false
          },
          "dp_lock_enabled": {
            "type": "boolean",
            "title": "Dp Lock Enabled",
            "default": false
          },
          "audio_output_mode": {
            "type": "string",
            "enum": [
              "pc",
              "phone",
              "both"
            ],
            "title": "Audio Output Mode",
            "default": "pc"
          },
          "audio_sync_offset_ms": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": -300.0,
            "title": "Audio Sync Offset Ms",
            "default": 0
          },
          "enable_flex_display": {
            "type": "boolean",
            "title": "Enable Flex Display",
            "default": false
          },
          "video_codec": {
            "type": "string",
            "enum": [
              "auto",
              "h265",
              "h264",
              "av1"
            ],
            "title": "Video Codec",
            "default": "auto"
          },
          "pixel_perfect_dpr": {
            "type": "boolean",
            "title": "Pixel Perfect Dpr",
            "default": true
          },
          "resize_instant_apply": {
            "type": "boolean",
            "title": "Resize Instant Apply",
            "default": false
          },
          "sharpening_mode": {
            "type": "string",
            "enum": [
              "off",
              "adaptive",
              "ultra"
            ],
            "title": "Sharpening Mode",
            "default": "adaptive"
          },
          "stealth_dpi_enabled": {
            "type": "boolean",
            "title": "Stealth Dpi Enabled",
            "default": true
          },
          "density_refresh_enabled": {
            "type": "boolean",
            "title": "Density Refresh Enabled",
            "default": true
          },
          "density_refresh_hard": {
            "type": "boolean",
            "title": "Density Refresh Hard",
            "default": false
          },
          "handoff_prelanding": {
            "type": "boolean",
            "title": "Handoff Prelanding",
            "default": true
          },
          "windowing_mode": {
            "type": "string",
            "enum": [
              "eco",
              "independent",
              "hybrid_auto"
            ],
            "title": "Windowing Mode",
            "default": "hybrid_auto"
          },
          "remember_devices": {
            "type": "boolean",
            "title": "Remember Devices",
            "default": true
          },
          "phone_handoff_windowing": {
            "type": "string",
            "enum": [
              "fullscreen",
              "freeform"
            ],
            "title": "Phone Handoff Windowing",
            "default": "fullscreen"
          },
          "app_closed_behavior": {
            "type": "string",
            "enum": [
              "close",
              "badge"
            ],
            "title": "App Closed Behavior",
            "default": "close"
          }
        },
        "type": "object",
        "title": "ProjectSettings"
      },
      "QrPayload": {
        "properties": {
          "service_name": {
            "type": "string",
            "title": "Service Name"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "service_name",
          "password",
          "text"
        ],
        "title": "QrPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The per-user API token (`OPENDEX_API_TOKEN` or `~/.opendex/api-token`)."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "The same token as a query parameter — accepted on GET requests and WebSocket handshakes only."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "The API token is missing or wrong (WebSocket: close code 4401).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "ForeignOrigin": {
        "description": "The request carries an `Origin` that is not one of the app's own origins, or a `Host` that is not a loopback name (WebSocket: close code 4403).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "LengthRequired": {
        "description": "The request body was sent with chunked transfer-encoding and no `Content-Length` (chunked bodies are refused).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
      "PayloadTooLarge": {
        "description": "The request body is larger than `OPENDEX_MAX_BODY_BYTES` (default 2 MB).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8710",
      "description": "Default local listener (OPENDEX_HTTP_HOST / OPENDEX_HTTP_PORT)"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "tokenQuery": []
    }
  ]
}